apps/orderbook_viewer.py
#!/usr/bin/env python3
"""
Real-time Orderbook Terminal UI for Polymarket 15-minute markets.
Read-only monitoring tool. No trades are executed.

Usage:
    python apps/orderbook_viewer.py --coin ETH
    python apps/orderbook_viewer.py --coin BTC --levels 10
"""

import os
import sys
import asyncio
import argparse
import logging
from pathlib import Path

logging.getLogger("src.websocket_client").setLevel(logging.WARNING)

from dotenv import load_dotenv
load_dotenv()

sys.path.insert(0, str(Path(__file__).parent.parent))

from lib import MarketManager, PriceTracker, Colors
from lib.terminal_utils import format_countdown


class OrderbookTUI:
    def __init__(self, coin: str = "ETH"):
        self.coin = coin.upper()
        self.market = MarketManager(coin=self.coin)
        self.prices = PriceTracker()
        self.running = False

    async def run(self) -> None:
        self.running = True

        @self.market.on_book_update
        async def handle_book(snapshot):
            for side, token_id in self.market.token_ids.items():
                if token_id == snapshot.asset_id:
                    self.prices.record(side, snapshot.mid_price)
                    break

        @self.market.on_connect
        def on_connect():
            pass

        @self.market.on_disconnect
        def on_disconnect():
            pass

        if not await self.market.start():
            print(f"{Colors.RED}Failed to start market manager{Colors.RESET}")
            return

        await self.market.wait_for_data(timeout=5.0)

        try:
            while self.running:
                self.render()
                await asyncio.sleep(0.1)
        except KeyboardInterrupt:
            pass
        finally:
            await self.market.stop()

    def render(self) -> None:
        lines = []
        ws_status = f"{Colors.GREEN}Connected{Colors.RESET}" if self.market.is_connected else f"{Colors.RED}Disconnected{Colors.RESET}"
        market = self.market.current_market
        countdown = "--:--"
        if market:
            mins, secs = market.get_countdown()
            countdown = format_countdown(mins, secs)

        lines.append(f"{Colors.BOLD}{'='*80}{Colors.RESET}")
        lines.append(f"{Colors.CYAN}Orderbook TUI{Colors.RESET} | {self.coin} | {ws_status} | Ends: {countdown}")
        lines.append(f"{Colors.BOLD}{'='*80}{Colors.RESET}")

        if market:
            lines.append(f"Market: {market.question}")
            lines.append(f"Slug: {market.slug}")
            lines.append("")

        up_ob = self.market.get_orderbook("up")
        down_ob = self.market.get_orderbook("down")

        lines.append(f"{Colors.GREEN}{'UP':^39}{Colors.RESET}|{Colors.RED}{'DOWN':^39}{Colors.RESET}")
        lines.append(f"{'Bid':>9} {'Size':>9} | {'Ask':>9} {'Size':>9}|{'Bid':>9} {'Size':>9} | {'Ask':>9} {'Size':>9}")
        lines.append("-" * 80)

        up_bids = up_ob.bids[:10] if up_ob else []
        up_asks = up_ob.asks[:10] if up_ob else []
        down_bids = down_ob.bids[:10] if down_ob else []
        down_asks = down_ob.asks[:10] if down_ob else []

        for i in range(10):
            up_bid = f"{up_bids[i].price:>9.4f} {up_bids[i].size:>9.1f}" if i < len(up_bids) else f"{'--':>9} {'--':>9}"
            up_ask = f"{up_asks[i].price:>9.4f} {up_asks[i].size:>9.1f}" if i < len(up_asks) else f"{'--':>9} {'--':>9}"
            down_bid = f"{down_bids[i].price:>9.4f} {down_bids[i].size:>9.1f}" if i < len(down_bids) else f"{'--':>9} {'--':>9}"
            down_ask = f"{down_asks[i].price:>9.4f} {down_asks[i].size:>9.1f}" if i < len(down_asks) else f"{'--':>9} {'--':>9}"
            lines.append(f"{up_bid} | {up_ask}|{down_bid} | {down_ask}")

        lines.append("-" * 80)

        up_mid = up_ob.mid_price if up_ob else 0
        down_mid = down_ob.mid_price if down_ob else 0
        up_spread = self.market.get_spread("up")
        down_spread = self.market.get_spread("down")

        lines.append(
            f"Mid: {Colors.GREEN}{up_mid:.4f}{Colors.RESET}  Spread: {up_spread:.4f}           |"
            f"Mid: {Colors.RED}{down_mid:.4f}{Colors.RESET}  Spread: {down_spread:.4f}"
        )

        up_history = self.prices.get_history_count("up")
        down_history = self.prices.get_history_count("down")
        up_vol = self.prices.get_volatility("up", 60)
        down_vol = self.prices.get_volatility("down", 60)

        lines.append("")
        lines.append(f"History: UP={up_history} DOWN={down_history} | 60s Volatility: UP={up_vol:.4f} DOWN={down_vol:.4f}")
        lines.append(f"{Colors.BOLD}{'='*80}{Colors.RESET}")
        lines.append(f"{Colors.DIM}Press Ctrl+C to exit{Colors.RESET}")

        output = "\033[H\033[J" + "\n".join(lines)
        print(output, flush=True)


def main():
    parser = argparse.ArgumentParser(description="Orderbook TUI for Polymarket 15-minute markets")
    parser.add_argument("--coin", type=str, default="ETH", choices=["BTC", "ETH", "SOL", "XRP"], help="Coin to monitor (default: ETH)")
    args = parser.parse_args()
    tui = OrderbookTUI(coin=args.coin)
    try:
        asyncio.run(tui.run())
    except KeyboardInterrupt:
        print("\nExiting...")


if __name__ == "__main__":
    main()
